misc: add / for forward compatibility of the lockfile
