+++ crates/rattler_conda_types/src/utils/url.rs
@@ -1,3 +1,6 @@
+use std::borrow::Cow;
+use url::Url;
+
 /// Parses the schema part of the human-readable channel. Returns the scheme part if it exists.
 pub(crate) fn parse_scheme(channel: &str) -> Option<&str> {
     let scheme_end = channel.find("://")?;
@@ -20,5 +23,16 @@ pub(crate) fn parse_scheme(channel: &str) -> Option<&str> {
         Some(scheme_part)
     } else {
         None
+    }
+}
+
+pub(crate) fn add_trailing_slash(url: &Url) -> Cow<'_, Url> {
+    let path = url.path();
+    if path.ends_with('/') {
+        Cow::Borrowed(url)
+    } else {
+        let mut url = url.clone();
+        url.set_path(&format!("{path}/"));
+        Cow::Owned(url)
     }
 }
